Fix the reentrance of the github issue tests
right now they panic
with this change they won't
serve cold
enjoy!
