+++ test/mako/alerter/github/issue.go
@@ -90,56 +90,6 @@ func Setup(org, repo, githubTokenPath string, dryrun bool) (*IssueHandler, error
 	return &IssueHandler{client: ghc, config: conf}, nil
 }
 
-// CreateIssueForTest will try to add an issue with the given testName and description.
-// If there is already an issue related to the test, it will try to update that issue.
-func (gih *IssueHandler) CreateIssueForTest(testName, desc string) error {
-	title := fmt.Sprintf(issueTitleTemplate, testName)
-	issue, err := gih.findIssue(title)
-	if err != nil {
-		return fmt.Errorf("failed to find issues for test %q: %w, skipped creating new issue", testName, err)
-	}
-	// If the issue hasn't been created, create one
-	if issue == nil {
-		commentBody := fmt.Sprintf(issueBodyTemplate, gih.config.repo, testName)
-		issue, err := gih.createNewIssue(title, commentBody)
-		if err != nil {
-			return fmt.Errorf("failed to create a new issue for test %q: %w", testName, err)
-		}
-		commentBody = fmt.Sprintf(issueSummaryCommentTemplate, desc)
-		if err := gih.addComment(*issue.Number, commentBody); err != nil {
-			return fmt.Errorf("failed to add comment for new issue %d: %w", *issue.Number, err)
-		}
-		return nil
-	}
-
-	// If the issue has been created, edit it
-	issueNumber := *issue.Number
-
-	// If the issue has been closed, reopen it
-	if *issue.State == string(ghutil.IssueCloseState) {
-		if err := gih.reopenIssue(issueNumber); err != nil {
-			return fmt.Errorf("failed to reopen issue %d: %w", issueNumber, err)
-		}
-		if err := gih.addComment(issueNumber, reopenIssueComment); err != nil {
-			return fmt.Errorf("failed to add comment for reopened issue %d: %w", issueNumber, err)
-		}
-	}
-
-	// Edit the old comment
-	comments, err := gih.getComments(issueNumber)
-	if err != nil {
-		return fmt.Errorf("failed to get comments from issue %d: %w", issueNumber, err)
-	}
-	if len(comments) < 1 {
-		return fmt.Errorf("existing issue %d is malformed, cannot update", issueNumber)
-	}
-	commentBody := fmt.Sprintf(issueSummaryCommentTemplate, desc)
-	if err := gih.editComment(issueNumber, *comments[0].ID, commentBody); err != nil {
-		return fmt.Errorf("failed to edit the comment for issue %d: %w", issueNumber, err)
-	}
-	return nil
-}
-
 // createNewIssue will create a new issue, and add perfLabel for it.
 func (gih *IssueHandler) createNewIssue(title, body string) (*github.Issue, error) {
 	var newIssue *github.Issue
